cdcc.c: Cast time_t to long before formatting with %ld.
Fixes platforms where time_t and long are not the same size, like netbsd i386.


git-svn-id: svn://svn.code.sf.net/p/bitchx/code/trunk@473 13b04d17-f746-0410-82c6-800466cd88b0

+++ source/cdcc.c
@@ -356,7 +356,10 @@ static int r_info(char *args, char *rest)
 			if (matchmcommand(q, ptr->num))
 				break;
 		if (ptr)
-			queue_send_to_server(from_server, "NOTICE %s :%d file%s %d gets %ld size %2.4f minspeed %ld time added",args, ptr->numfiles, plural(ptr->numfiles), ptr->gets, ptr->size, ptr->minspeed, ptr->timeadded); 
+			queue_send_to_server(from_server, 
+				"NOTICE %s :%d file%s %d gets %ld size %2.4f minspeed %ld time added",
+				args, ptr->numfiles, plural(ptr->numfiles), ptr->gets, 
+				ptr->size, ptr->minspeed, (long)ptr->timeadded); 
 		else
 			queue_send_to_server(from_server, "NOTICE %s :Invalid info request", args);
 	}
